Latest updates at 08:40 hours MDT on 06/12/2024

+++ xyLOGIX.Core.Debug/PostSharpLoggingInfrastructure.cs
@@ -249,8 +249,9 @@ namespace xyLOGIX.Core.Debug
                 );
 
                 PrepareLogFile(_relay);
-                
-                DeleteLogIfExists();
+
+                if (overwrite)
+                    DeleteLogIfExists();
                 
                 WriteTimestamp();
             }
